+++ app/src/main/java/chilladvanced/NativeLibraryTracker.kt
@@ -1,0 +1,13 @@
+package chilladvanced
+
+class NativeLibraryTracker {
+
+
+    fun start() {
+
+    }
+
+    fun stop() {
+
+    }
+}
