+++ app/src/main/java/chilladvanced/com/google/android/exoplayer2/ExoPlayerWithStat.kt
@@ -1,16 +1,12 @@
-package chilladvanced
+package chilladvanced.com.google.android.exoplayer2
 
-import android.net.Uri
+import chilladvanced.Logger
+import chilladvanced.NETWORK_METHODS
 import com.google.android.exoplayer2.ExoPlayer
 import com.google.android.exoplayer2.MediaItem
-import com.google.android.exoplayer2.source.MediaSource
-
 
 class ExoPlayerWithStat(val exoPlayer: ExoPlayer) : ExoPlayer by exoPlayer {
     override fun prepare() {
-
-
-
         exoPlayer.prepare()
         val mediaSource: MediaItem? = exoPlayer.currentMediaItem
         Logger.register(
